Add authorization for tasks management
This adds authorieation for managing tasks. A user can only update and delete their own tasks but an administrator can update and delete every task including anonymous tasks.

Resolves: #6.

+++ src/Controller/TaskController.php
@@ -7,15 +7,25 @@ use App\Form\TaskType;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Security\Core\Security;
 
 class TaskController extends Controller
 {
+    private $security;
+
+    public function __construct(Security $security)
+    {
+        $this->security = $security;
+    }
+
     /**
      * @Route("/tasks", name="task_list")
      */
     public function listAction()
     {
-        return $this->render('task/list.html.twig', ['tasks' => $this->getDoctrine()->getRepository('App:Task')->findAll()]);
+        $tasks = $this->getDoctrine()->getRepository(Task::class)->findBy(['user' => $this->getUser()]);
+
+        return $this->render('task/list.html.twig', ['tasks' => $tasks]);
     }
 
     /**
@@ -49,6 +59,10 @@ class TaskController extends Controller
      */
     public function editAction(Task $task, Request $request)
     {
+        if (!$this->security->isGranted('ROLE_ADMIN') && !$this->security->isGranted('UPDATE', $task)) {
+            throw $this->createAccessDeniedException();
+        }
+
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
@@ -72,6 +86,10 @@ class TaskController extends Controller
      */
     public function toggleTaskAction(Task $task)
     {
+        if (!$this->security->isGranted('ROLE_ADMIN') && !$this->security->isGranted('TOGGLE', $task)) {
+            throw $this->createAccessDeniedException();
+        }
+
         $task->toggle(!$task->isDone());
         $this->getDoctrine()->getManager()->flush();
 
@@ -85,6 +103,10 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
+        if (!$this->security->isGranted('ROLE_ADMIN') && !$this->security->isGranted('DELETE', $task)) {
+            throw $this->createAccessDeniedException();
+        }
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($task);
         $em->flush();
